\name{rucl2-package}
\alias{rucl2-package}
\alias{rucl2}
\docType{package}
\title{
\packageTitle{rucl2}
}
\description{
\packageDescription{rucl2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rucl2}
\packageIndices{rucl2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rucl2}

Maintainer: \packageMaintainer{rucl2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
